Made ArcaneTheme private (_ArcaneTheme)
