+++ lib/src/services/reactive_theme/reactive_theme_extensions.dart
@@ -19,11 +19,3 @@ extension DarkMode on BuildContext {
     return brightness == Brightness.dark;
   }
 }
-
-extension ArcaneThemeContext on BuildContext {
-  /// Get the current theme mode from the nearest ArcaneThemeInherited widget
-  ThemeMode get themeMode {
-    return ArcaneTheme.of(this)?.themeMode ??
-        ArcaneReactiveTheme.I.currentThemeMode;
-  }
-}
